v2.2.0 release ledger sync — critical-eval-driven improvement bundle
Implements P0/P1 items from the v2.2.0 critical evaluation of the
v2.1.0 baseline. Codex's v2.0.0 pass had already closed identity
attribution + ROI loop + anomaly timestamp normalization, so this
release is pruning-first: dead signals removed, double-firing rules
deduplicated, magic thresholds tagged with provenance, and the
anomaly subsystem made honest about per-detector provider asymmetry.

P0 (dead-code purge + dedup):
- Drop SignalSet.repeated_output field — never populated by any
  adapter; alert + advisory rules consuming it were silent dead
  paths. Removed alongside 13 test references.
- Drop TaskType::{CodeExploration, LogTriage, Summary} variants —
  no adapter populated them. Kept Review, Automation, SessionResume.
  code_exploration advisory now triggers solely on verbose_answer
  (itself flagged for measurement-deprecation in VERBOSE_MARKERS
  docstring).
- log_storm was firing twice (Warning alert + Concern advisory with
  quota_tight aggressive variant). Alert path removed — advisory
  retained with the runnable archive command + escalation.
- quota_pressure_recommendations collapsed from three windows
  firing simultaneously (generic + 5h + weekly) to one most-urgent
  rec selected by smallest critical_pct - value headroom, with
  5h > weekly > generic tiebreak. Two new tests pin both
  directions.

P0-3 (threshold provenance):
- Every magic number in PolicyGates::default() now carries an
  inline measured/intuition/mirrors/safety tag. config/qmonster.
  example.toml [anomaly] header points operators at the inline
  source-side tags.

P1-1 (detector × provider eligibility):
- AnomalyKind::supported_providers() declares per-detector adapter
  coverage. eval_anomalies takes a Provider parameter and short-
  circuits ineligible detectors before they run. Closes the
  MemoryGrowth-silent-on-Claude honesty gap surfaced by the
  critical eval. Settings Rules tab renders ` · supports: …`
  chips on asymmetric detector rows.

P1-2 (SubagentSideEffect honesty):
- promote_anomalies_to_recommendations reason leads with
  '⚠ correlation only — providers do not expose per-subagent token
  attribution.' Detector emits a cooccurring_kinds evidence row
  alongside the renamed subagent_correlation_NOT_attribution metric.

P1-3 (recommendation acknowledgement):
- RecommendationOutcome::Copied variant added. New
  copy_selected_alert_command_to_clipboard_with_ledger writes the
  outcome on successful y-copy via parse_recommendation_key. Ledger
  failure does not surface to the operator.

P2 deferrals (.mission/decisions/MDR-DRAFT-v2.2.0-*.md):
- P2-1 anomaly subsystem keep-or-trim awaits 2 weeks of P1-3 ack
  telemetry.
- P2-3 severity/source enum simplification awaits operator
  approval — contract-change cost vs cognitive-load reduction
  trade-off documented.
- P2-2 overlay consolidation deliberately excluded per operator
  instruction during the v2.2.0 pass.

Tests: 1491 lib + 136 integration, all green. fmt + clippy clean.

+++ src/app/tui_loop.rs
@@ -991,8 +991,7 @@ where
                                         // `y` copy lands as a
                                         // `RecommendationOutcome::Copied` row
                                         // when the alert is recommendation-class.
-                                        let now_unix_ms =
-                                            crate::app::event_loop::current_unix_ms() as i64;
+                                        let now_unix_ms = crate::app::event_loop::current_unix_ms();
                                         let notice =
                                             copy_selected_alert_command_to_clipboard_with_ledger(
                                                 view,
